requirements.txt
imap-tools==1.10.0
python-dotenv==1.1.0
schedule==1.2.2
